ecoloop-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
aiofiles==24.1.0
httpx==0.27.2
pydantic[email]==2.9.2
email-validator==2.2.0
python-dotenv==1.0.1

# ── PostgreSQL support ──────────────────────────────
psycopg2-binary==2.9.9

# ── Email notifications ─────────────────────────────
sendgrid==6.11.0

# ── WhatsApp (Twilio) ───────────────────────────────
twilio==9.3.6

# ── Razorpay ────────────────────────────────────────
razorpay==1.4.1

# ── Async email via SMTP ────────────────────────────
aiosmtplib==3.0.1
